Major rework to the script
Removed the first line relating to the alias due to the script functioning fine without it. Added a full path to the 'git' executable in the config line, which fixed the "command not found" error when executing as a script. Code now works as intended in both script and manual executions. Added comments for each line.